+++ blueprints/flexberry-core/files/__root__/controllers/application.js
@@ -11,13 +11,7 @@ export default Ember.Controller.extend({
           caption: i18n.t('forms.application.sitemap.index.caption'),
           title: i18n.t('forms.application.sitemap.index.title'),
           children: null
-        },
-        {
-          link: null,
-          caption: i18n.t('forms.application.sitemap.application.caption'),
-          title: i18n.t('forms.application.sitemap.application.title'),
-          children: [<%=children%>]
-        }
+        },<%=children%>
       ]
     };
   }),