+++ blueprints/flexberry-core/index.js
@@ -86,7 +86,7 @@ var CoreBlueprint = (function () {
         for (var _c = 0, _d = this.sitemap.items; _c < _d.length; _c++) {
             var item = _d[_c];
             var childItemExt = new SitemapItemExt(item);
-            childItemExt.process("forms.application.sitemap.application", 2);
+            childItemExt.process("forms.application.sitemap", 5);
             applicationMenuLocales.push(childItemExt.translation, childItemExt.translationOtherLocales);
             children.push(childItemExt.sitemap);
         }
@@ -143,9 +143,9 @@ var SitemapItemExt = (function () {
         }
         this.translation = ("" + indentStr2 + this.quote(translationName) + ": {\n" + indentStr + "caption: '" + this.escapeValue(this.baseItem.caption) + "',\n") +
             (indentStr + "title: '" + this.escapeValue(this.baseItem.title) + "',\n" + childrenStr + "\n" + indentStr2 + "}");
-        this.translationOtherLocales = (indentStr2 + "'" + this.quote(translationName) + "': {\n" + indentStr + "caption: '" + this.escapeValue(translationName) + "',\n") +
+        this.translationOtherLocales = ("" + indentStr2 + this.quote(translationName) + ": {\n" + indentStr + "caption: '" + this.escapeValue(translationName) + "',\n") +
             (indentStr + "title: '" + this.escapeValue(translationName) + "',\n" + childrenOtherLocalesStr + "\n" + indentStr2 + "}");
-        var INDENT = "        ";
+        var INDENT = "";
         this.sitemap = ("{\n" + INDENT + indentStr + "link: " + this.quoteIfNotNull(this.baseItem.link) + ",\n") +
             ("" + INDENT + indentStr + "caption: i18n.t('" + translationProp + ".caption'),\n") +
             ("" + INDENT + indentStr + "title: i18n.t('" + translationProp + ".title'),\n") +