+++ tests/dummy/app/controllers/components-examples/flexberry-dropdown/settings-example.js
@@ -38,7 +38,7 @@ export default Ember.Controller.extend({
    */
   componentTemplateText: new Ember.Handlebars.SafeString(
     '{{flexberry-dropdown<br>' +
-    '  items=(enum-captions \"components-examples/flexberry-dropdown/settings-example/enumeration\")<br>' +
+    '  items=(flexberry-enum \"components-examples/flexberry-dropdown/settings-example/enumeration\")<br>' +
     '  value=model.enumeration<br>' +
     '  placeholder=placeholder<br>' +
     '  readonly=readonly<br>' +