+++ CHANGELOG.md
@@ -5,9 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 * Transforms:
-    * Moved transforms to ember-flexberry-data addon.
+    * Moved transforms to [`ember-flexberry-data`](https://github.com/Flexberry/ember-flexberry-data) addon.
 * Enums:
-    * Moved enum initializers to ember-flexberry-data addon.
+    * Moved enum initializers to [`ember-flexberry-data`](https://github.com/Flexberry/ember-flexberry-data) addon.
 
 ## [0.3.0] - 2016-07-29
 ### Added